Change: theme-ified add travel page
