+++ components/modal/addModal/AddLapModal.tsx
@@ -69,7 +69,7 @@ export default function AddLapModal({ stops, isModalVisible, onClose, onSelect }
                         <View style={[modalStyles.modalContainer, modalStyles.lapModalContainer]}>
                             <View style={inputElementStyles['light'].inputGroup}>
                                 <Text style={inputElementStyles['light'].inputLabel}>Time:</Text>
-                                <Pressable onPress={() => setShowDatetimePicker(true)} style={inputStyles.pressableInput}>
+                                <Pressable onPress={() => setShowDatetimePicker(true)} style={inputStyles['light'].pressableInput}>
                                     <Text style={inputElementStyles['light'].insideLabel}>{formatDateForDisplay(lap.time)}</Text>
                                 </Pressable>
                             </View>
@@ -77,7 +77,7 @@ export default function AddLapModal({ stops, isModalVisible, onClose, onSelect }
                             <View style={inputElementStyles['light'].inputGroup}>
                                 <Text style={inputElementStyles['light'].inputLabel}>Stop:</Text>
                                 <Pressable
-                                    style={inputStyles.pressableInput}
+                                    style={inputStyles['light'].pressableInput}
                                     onPress={() => openStopModal('last_stop_id')}>
                                     <Text style={[inputElementStyles['light'].insideLabel, { marginBottom: 0 }]}>
                                         {stops.find(item => item.id === lap.stop_id)?.name || 'Select Stop'}
@@ -104,7 +104,7 @@ export default function AddLapModal({ stops, isModalVisible, onClose, onSelect }
                                     returnKeyType="done"
                                     multiline={true}
                                     numberOfLines={3}
-                                    style={[inputStyles.textInput, inputStyles.multilineTextInput, inputElementStyles['light'].insideLabel]}
+                                    style={[inputStyles['light'].textInput, inputStyles['light'].multilineTextInput, inputElementStyles['light'].insideLabel]}
                                 />
                             </View>
