+++ components/modal/addModal/AddVehicleTypeModal.tsx
@@ -34,7 +34,7 @@ export default function AddVehicleTypeModal({ icons, onSubmit, onCancel }: Vehic
                         <View style={inputElementStyles['light'].inputGroup}>
                             <Text style={inputElementStyles['light'].insideLabel}>Name:</Text>
                             <TextInput
-                                style={inputStyles.pressableInput}
+                                style={inputStyles['light'].pressableInput}
                                 placeholder="e.g., Standard Bus"
                                 value={vehicleType.name}
                                 onChangeText={text => setVehicleType({ ...vehicleType, "name": text })}
