+++ components/modal/editModal/EditDirectionModal.tsx
@@ -34,7 +34,7 @@ export default function EditDirectionModal({ onCancel, onSubmit }: BaseModalCont
                         <View style={inputElementStyles['light'].inputGroup}>
                             <Text style={inputElementStyles['light'].insideLabel}>Name:</Text>
                             <TextInput
-                                style={inputStyles.pressableInput}
+                                style={inputStyles['light'].pressableInput}
                                 placeholder="Direction name..."
                                 value={direction.name}
                                 onChangeText={text => (setDirection({ ...direction, "name": text }))}
