+++ const/color.ts
@@ -8,5 +8,6 @@ export const colors = {
     },
     text: {
         dimWhite: '#e8e8e8',
+        dimmerWhite: '#b3b3b3',
     }
 }
